refactor: boardStore 관계 추가 및 리팩토링
